move g.getBaseUrl() to a separate file creds2.go and test to creds2_test.go to minimize changing creds.go
