+++ ext/git/creds.go
@@ -42,7 +42,6 @@ const (
 	// githubAccessTokenUsername is a username that is used to with the github access token
 	githubAccessTokenUsername = "x-access-token"
 	forceBasicAuthHeaderEnv   = "ARGOCD_GIT_AUTH_HEADER"
-	defaultGithubApiUrl       = "https://api.github.com"
 )
 
 func init() {
@@ -462,27 +461,6 @@ func (g GitHubAppCreds) getAccessToken() (string, error) {
 	githubAppTokenCache.Set(key, itr, time.Minute*60)
 
 	return itr.Token(ctx)
-}
-
-func (g GitHubAppCreds) getBaseURL() string {
-	if g.baseURL != "" {
-		return strings.TrimSuffix(g.baseURL, "/")
-	}
-	if g.repoURL == "" {
-		return defaultGithubApiUrl
-	}
-
-	repoUrl, err := url.Parse(g.repoURL)
-	if err != nil || repoUrl.Hostname() == "github.com" {
-		return defaultGithubApiUrl
-	}
-
-	// GitHub Enterprise
-	scheme := repoUrl.Scheme
-	if scheme == "" {
-		scheme = "https"
-	}
-	return fmt.Sprintf("%s://%s/api/v3", scheme, repoUrl.Host)
 }
 
 func (g GitHubAppCreds) HasClientCert() bool {
